Brushing off the Dust (#24)
Some things have changed with the way Dune returns things:

    Response bytes can be bigger than u16
    Floats are returned as strings
    Timestamps no longer have the T in the middle

Most of these things are since the move from Spark to Trino (since this project has not been touched since then).

Furthermore we

    Updated dependencies
    Fixed dead reference links
    Migrated an old query to Trino
    yada yada yada..

Nobody even watches this repo, none of this matters to anyone but me... But it will matter.

+++ src/response.rs
@@ -1,4 +1,4 @@
-use crate::dateutil::{datetime_from_str, optional_datetime_from_str};
+use crate::parse_utils::{datetime_from_str, optional_datetime_from_str};
 use chrono::{DateTime, Utc};
 use serde::Deserialize;
 use serde_with::DeserializeFromStr;
@@ -66,7 +66,7 @@ pub struct CancellationResponse {
 #[derive(Deserialize, Debug)]
 pub struct ResultMetaData {
     pub column_names: Vec<String>,
-    pub result_set_bytes: u16,
+    pub result_set_bytes: u64,
     pub total_row_count: u32,
     pub datapoint_count: u32,
     pub pending_time_millis: Option<u32>,
